backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
gunicorn==23.0.0
python-dotenv==1.0.1
pydantic==2.10.6
supabase==2.10.0
groq==0.9.0
anthropic==0.28.0
requests==2.32.3
stripe==10.13.0
python-multipart==0.0.9
httpx==0.27.0
passlib==1.7.4
PyJWT==2.8.0
amadeus==10.1.0
reportlab==4.1.0